+++ dnsimple/resource_dnsimple_record.go
@@ -1,13 +1,14 @@
 package dnsimple
 
 import (
+	"context"
 	"fmt"
 	"log"
 	"strconv"
 	"strings"
 
 	"github.com/dnsimple/dnsimple-go/dnsimple"
-	"github.com/hashicorp/terraform/helper/schema"
+	"github.com/hashicorp/terraform-plugin-sdk/helper/schema"
 )
 
 func resourceDNSimpleRecord() *schema.Resource {
@@ -72,27 +73,26 @@ func resourceDNSimpleRecordCreate(d *schema.ResourceData, meta interface{}) erro
 	provider := meta.(*Client)
 
 	// Create the new record
-	newRecord := dnsimple.ZoneRecord{
-		Name:    d.Get("name").(string),
+	recordAttributes := dnsimple.ZoneRecordAttributes{
+		Name:    dnsimple.String(d.Get("name").(string)),
 		Type:    d.Get("type").(string),
 		Content: d.Get("value").(string),
 	}
 	if attr, ok := d.GetOk("ttl"); ok {
-		newRecord.TTL, _ = strconv.Atoi(attr.(string))
+		recordAttributes.TTL, _ = strconv.Atoi(attr.(string))
 	}
-
 	if attr, ok := d.GetOk("priority"); ok {
-		newRecord.Priority, _ = strconv.Atoi(attr.(string))
+		recordAttributes.Priority, _ = strconv.Atoi(attr.(string))
 	}
 
-	log.Printf("[DEBUG] DNSimple Record create configuration: %#v", newRecord)
+	log.Printf("[DEBUG] DNSimple Record create recordAttributes: %#v", recordAttributes)
 
-	resp, err := provider.client.Zones.CreateRecord(provider.config.Account, d.Get("domain").(string), newRecord)
+	resp, err := provider.client.Zones.CreateRecord(context.Background(), provider.config.Account, d.Get("domain").(string), recordAttributes)
 	if err != nil {
 		return fmt.Errorf("Failed to create DNSimple Record: %s", err)
 	}
 
-	d.SetId(strconv.Itoa(resp.Data.ID))
+	d.SetId(strconv.FormatInt(resp.Data.ID, 10))
 	log.Printf("[INFO] DNSimple Record ID: %s", d.Id())
 
 	return resourceDNSimpleRecordRead(d, meta)
@@ -101,14 +101,14 @@ func resourceDNSimpleRecordCreate(d *schema.ResourceData, meta interface{}) erro
 func resourceDNSimpleRecordRead(d *schema.ResourceData, meta interface{}) error {
 	provider := meta.(*Client)
 
-	recordID, err := strconv.Atoi(d.Id())
+	recordID, err := strconv.ParseInt(d.Id(), 10, 64)
 	if err != nil {
 		return fmt.Errorf("Error converting Record ID: %s", err)
 	}
 
-	resp, err := provider.client.Zones.GetRecord(provider.config.Account, d.Get("domain").(string), recordID)
+	resp, err := provider.client.Zones.GetRecord(context.Background(), provider.config.Account, d.Get("domain").(string), recordID)
 	if err != nil {
-		if err != nil && strings.Contains(err.Error(), "404") {
+		if strings.Contains(err.Error(), "404") {
 			log.Printf("DNSimple Record Not Found - Refreshing from State")
 			d.SetId("")
 			return nil
@@ -136,33 +136,31 @@ func resourceDNSimpleRecordRead(d *schema.ResourceData, meta interface{}) error 
 func resourceDNSimpleRecordUpdate(d *schema.ResourceData, meta interface{}) error {
 	provider := meta.(*Client)
 
-	recordID, err := strconv.Atoi(d.Id())
+	recordID, err := strconv.ParseInt(d.Id(), 10, 64)
 	if err != nil {
 		return fmt.Errorf("Error converting Record ID: %s", err)
 	}
 
-	updateRecord := dnsimple.ZoneRecord{}
-
+	recordAttributes := dnsimple.ZoneRecordAttributes{}
 	if attr, ok := d.GetOk("name"); ok {
-		updateRecord.Name = attr.(string)
+		recordAttributes.Name = dnsimple.String(attr.(string))
 	}
 	if attr, ok := d.GetOk("type"); ok {
-		updateRecord.Type = attr.(string)
+		recordAttributes.Type = attr.(string)
 	}
 	if attr, ok := d.GetOk("value"); ok {
-		updateRecord.Content = attr.(string)
+		recordAttributes.Content = attr.(string)
 	}
 	if attr, ok := d.GetOk("ttl"); ok {
-		updateRecord.TTL, _ = strconv.Atoi(attr.(string))
+		recordAttributes.TTL, _ = strconv.Atoi(attr.(string))
 	}
-
 	if attr, ok := d.GetOk("priority"); ok {
-		updateRecord.Priority, _ = strconv.Atoi(attr.(string))
+		recordAttributes.Priority, _ = strconv.Atoi(attr.(string))
 	}
 
-	log.Printf("[DEBUG] DNSimple Record update configuration: %#v", updateRecord)
+	log.Printf("[DEBUG] DNSimple Record update configuration: %#v", recordAttributes)
 
-	_, err = provider.client.Zones.UpdateRecord(provider.config.Account, d.Get("domain").(string), recordID, updateRecord)
+	_, err = provider.client.Zones.UpdateRecord(context.Background(), provider.config.Account, d.Get("domain").(string), recordID, recordAttributes)
 	if err != nil {
 		return fmt.Errorf("Failed to update DNSimple Record: %s", err)
 	}
@@ -175,12 +173,12 @@ func resourceDNSimpleRecordDelete(d *schema.ResourceData, meta interface{}) erro
 
 	log.Printf("[INFO] Deleting DNSimple Record: %s, %s", d.Get("domain").(string), d.Id())
 
-	recordID, err := strconv.Atoi(d.Id())
+	recordID, err := strconv.ParseInt(d.Id(), 10, 64)
 	if err != nil {
 		return fmt.Errorf("Error converting Record ID: %s", err)
 	}
 
-	_, err = provider.client.Zones.DeleteRecord(provider.config.Account, d.Get("domain").(string), recordID)
+	_, err = provider.client.Zones.DeleteRecord(context.Background(), provider.config.Account, d.Get("domain").(string), recordID)
 	if err != nil {
 		return fmt.Errorf("Error deleting DNSimple Record: %s", err)
 	}
@@ -192,7 +190,7 @@ func resourceDNSimpleRecordImport(d *schema.ResourceData, meta interface{}) ([]*
 	parts := strings.Split(d.Id(), "_")
 
 	if len(parts) != 2 {
-		return nil, fmt.Errorf("Error Importing dnsimple_record. Please make sure the record ID is in the form DOMAIN_RECORDID (i.e. example.com_1234")
+		return nil, fmt.Errorf("Error Importing dnsimple_record. Please make sure the record ID is in the form DOMAIN_RECORDID (i.e. example.com_1234)")
 	}
 
 	d.SetId(parts[1])